requirements.txt
networkx~=3.4.2
matplotlib~=3.10.0
numpy~=1.26.4
biopython~=1.84
scikit-learn~=1.5.1